requirements.txt
tkcalendar
matplotlib
psycopg2-binary
pytest
ruff
pyinstaller
tkcalendar
dotenv
